[ai-registry-adaptor] Rename mcp-registry module to ai-registry
