--- mcp-registry-adaptor/src/main/java/com/alibaba/nacos/mcpregistry/NacosMcpRegistry.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 1999-2025 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.mcpregistry;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.ldap.LdapAutoConfiguration;
-import org.springframework.context.annotation.PropertySource;
-
-/**
- * Nacos Mcp Registry.
- * This server implement the official mcp registry api.
- * @author xinluo
- */
-@SpringBootApplication(exclude = LdapAutoConfiguration.class)
-@PropertySource("classpath:nacos-mcp-registry.properties")
-public class NacosMcpRegistry {
-    
-    public static void main(String[] args) {
-        SpringApplication.run(NacosMcpRegistry.class, args);
-    }
-}
