--- mcp-registry-adaptor/src/main/java/com/alibaba/nacos/mcpregistry/form/SkillsFileQueryForm.java
@@ -1,80 +1,0 @@
-/*
- * Copyright 1999-2026 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.mcpregistry.form;
-
-import com.alibaba.nacos.api.exception.api.NacosApiException;
-import com.alibaba.nacos.api.model.NacosForm;
-import com.alibaba.nacos.api.model.v2.ErrorCode;
-import com.alibaba.nacos.common.utils.StringUtils;
-
-/**
- * Query form for skills CLI compatible file fetch endpoint.
- *
- * @author nacos
- */
-public class SkillsFileQueryForm implements NacosForm {
-
-    private String namespaceId;
-
-    private String skillName;
-
-    private String filePath;
-
-    @Override
-    public void validate() throws NacosApiException {
-        if (StringUtils.isBlank(namespaceId)) {
-            throw new NacosApiException(NacosApiException.INVALID_PARAM, ErrorCode.PARAMETER_MISSING,
-                    "namespaceId is required");
-        }
-        if (StringUtils.isBlank(skillName)) {
-            throw new NacosApiException(NacosApiException.INVALID_PARAM, ErrorCode.PARAMETER_MISSING,
-                    "skillName is required");
-        }
-        if (StringUtils.isBlank(filePath)) {
-            throw new NacosApiException(NacosApiException.INVALID_PARAM, ErrorCode.PARAMETER_MISSING,
-                    "filePath is required");
-        }
-        if (filePath.startsWith("/") || filePath.startsWith("\\") || filePath.contains("..")) {
-            throw new NacosApiException(NacosApiException.INVALID_PARAM, ErrorCode.PARAMETER_VALIDATE_ERROR,
-                    "filePath is invalid");
-        }
-    }
-
-    public String getNamespaceId() {
-        return namespaceId;
-    }
-
-    public void setNamespaceId(String namespaceId) {
-        this.namespaceId = namespaceId;
-    }
-
-    public String getSkillName() {
-        return skillName;
-    }
-
-    public void setSkillName(String skillName) {
-        this.skillName = skillName;
-    }
-
-    public String getFilePath() {
-        return filePath;
-    }
-
-    public void setFilePath(String filePath) {
-        this.filePath = filePath;
-    }
-}
