--- mcp-registry-adaptor/src/main/java/com/alibaba/nacos/mcpregistry/form/SkillsSearchForm.java
@@ -1,78 +1,0 @@
-/*
- * Copyright 1999-2026 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.mcpregistry.form;
-
-import com.alibaba.nacos.api.exception.api.NacosApiException;
-import com.alibaba.nacos.api.model.NacosForm;
-import com.alibaba.nacos.api.model.v2.ErrorCode;
-import com.alibaba.nacos.common.utils.StringUtils;
-
-/**
- * Query form for skills CLI compatible search endpoint.
- *
- * @author nacos
- */
-public class SkillsSearchForm implements NacosForm {
-
-    private static final int DEFAULT_LIMIT = 10;
-
-    private static final int MAX_LIMIT = 10;
-
-    private String namespaceId;
-
-    private String q;
-
-    private Integer limit;
-
-    @Override
-    public void validate() throws NacosApiException {
-        if (StringUtils.isBlank(namespaceId)) {
-            throw new NacosApiException(NacosApiException.INVALID_PARAM, ErrorCode.PARAMETER_MISSING,
-                    "namespaceId is required");
-        }
-        if (limit == null || limit <= 0) {
-            limit = DEFAULT_LIMIT;
-        }
-        if (limit > MAX_LIMIT) {
-            limit = MAX_LIMIT;
-        }
-    }
-
-    public String getNamespaceId() {
-        return namespaceId;
-    }
-
-    public void setNamespaceId(String namespaceId) {
-        this.namespaceId = namespaceId;
-    }
-
-    public String getQ() {
-        return q;
-    }
-
-    public void setQ(String q) {
-        this.q = q;
-    }
-
-    public Integer getLimit() {
-        return limit;
-    }
-
-    public void setLimit(Integer limit) {
-        this.limit = limit;
-    }
-}
