--- mcp-registry-adaptor/src/main/java/com/alibaba/nacos/mcpregistry/model/skills/SkillsSearchItem.java
@@ -1,65 +1,0 @@
-/*
- * Copyright 1999-2026 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.mcpregistry.model.skills;
-
-/**
- * Search result item for the skills CLI.
- *
- * @author nacos
- */
-public class SkillsSearchItem {
-
-    private String id;
-
-    private String name;
-
-    private long installs;
-
-    private String source;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public long getInstalls() {
-        return installs;
-    }
-
-    public void setInstalls(long installs) {
-        this.installs = installs;
-    }
-
-    public String getSource() {
-        return source;
-    }
-
-    public void setSource(String source) {
-        this.source = source;
-    }
-}
