--- mcp-registry-adaptor/src/main/java/com/alibaba/nacos/mcpregistry/model/skills/SkillsSearchResponse.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 1999-2026 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.mcpregistry.model.skills;
-
-import java.util.List;
-
-/**
- * Search response for the skills CLI.
- *
- * @author nacos
- */
-public class SkillsSearchResponse {
-
-    private List<SkillsSearchItem> skills;
-
-    public List<SkillsSearchItem> getSkills() {
-        return skills;
-    }
-
-    public void setSkills(List<SkillsSearchItem> skills) {
-        this.skills = skills;
-    }
-}
