--- mcp-registry-adaptor/src/main/resources/nacos-mcp-registry-banner.txt
@@ -1,14 +1,0 @@
-         ,--.                                           
-       ,--.'|                                           
-   ,--,:  : |                                           Nacos Mcp Registry ${application.version}
-,`--.'`|  ' :                       ,---.               Running in ${nacos.mode} mode, ${nacos.function.mode} function modules
-|   :  :  | |                      '   ,'\   .--.--.    Port: ${nacos.ai.mcp.registry.port}
-:   |   \ | :  ,--.--.     ,---.  /   /   | /  /    '   Pid: ${pid}
-|   : '  '; | /       \   /     \.   ; ,. :|  :  /`./   
-'   ' ;.    ;.--.  .-. | /    / ''   | |: :|  :  ;_     
-|   | | \   | \__\/: . ..    ' / '   | .; : \  \    `.        https://nacos.io
-'   : |  ; .' ," .--.; |'   ; :__|   :    |  `----.   \ 
-|   | '`--'  /  /  ,.  |'   | '.'|\   \  /  /  /`--'  / 
-'   : |     ;  :   .'   \   :    : `----'  '--'.     /  
-;   |.'     |  ,     .-./\   \  /            `--'---'   
-'---'        `--`---'     `----'
